fix code

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- function
 
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS FLOAT
+RETURNS INT
 
 BEGIN
     IF b = 0 THEN
